+++ frontend/src/components/utils/DropdownIndicator.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+import { Box } from "@chakra-ui/react";
+
+const DropdownIndicator = () => {
+  return (
+    <Box
+      borderLeft="5px solid transparent"
+      borderRight="5px solid transparent"
+      borderTop="9px solid black"
+      margin="10px"
+    />
+  );
+};
+
+export default DropdownIndicator;
